Build hyperion without bz2 dependency
The system-provided libbz2 shared object file is named/versioned
differently between Debian-type and RedHat-type Linux distributions,
it seems. We only use gzip compression for the DASD, not bz2 (and
bz2 use is rare in the Hercules community in general, it seems),
so no great loss to not support it in this build. One less thing
to go wrong when running the pre-build hyperion binary on different
systems.