Build hyperion without bz2 dependency... try 2
The system-provided libbz2 shared object file is named/versioned
differently between Debian-type and RedHat-type Linux distributions,
it seems. We only use gzip compression for the DASD, not bz2 (and
bz2 use is rare in the Hercules community in general, it seems),
so no great loss to not support it in this build. One less thing
to go wrong when running the pre-build hyperion binary on different
systems.

This second attempt builds on a Debian 11 system instead of Debian 10,
as Debian 10 ended up producing a build that was even less compatible
with a variety of newer distros. Debian 11 is what we originally built
with. It seems to be the current sweet spot of supporting slightly older
distros (Debian 11, Ubuntu 20.04, etc) but hopefully also
latest-and-greatest (Debian 12, Fedora 38, etc.)